Reorganized The Project Structure
Added a menu and instructions page.
The game is almost ready, just need to add the Goals and the point system (Hardest Part Of Them All) and change the Controls from VR to touchpad on the mobile device, as this will be exported to Android via Phonegap (Cordova). The application will have continual work.